+++ app/src/main/java/org/teamone/client/auth/UserController.java
@@ -1,0 +1,7 @@
+package org.teamone.client.auth;
+
+/**
+ * Created by daniel on 10/7/15.
+ */
+public class UserController {
+}
