+++ app/src/main/java/org/teamone/client/generic/User.java
@@ -1,0 +1,30 @@
+package org.teamone.client.generic;
+
+import java.util.Map;
+
+/**
+ * Created by daniel on 10/7/15.
+ */
+public class User {
+
+    private String message = "blahhlahsd";
+
+    public String username;
+    public String password;
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
