+++ app/src/main/webapp/WEB-INF/jsp/generic/navbar.jsp
@@ -1,4 +1,5 @@
 <%@ page contentType="text/html;charset=UTF-8" language="java" pageEncoding="UTF-8"%>
+<%@taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 
 <div id="navbar" class="navbar navbar-default" role="navigation">
   <div class="navbar-header">
@@ -12,6 +13,13 @@
   <div class="navbar-collapse collapse">
     <ul class="nav navbar-nav">
       <li><a href="../../../"> Home </a></li>
+      <c:forTokens items="Zara,nuha,login,navbar,login" delims="," var="name">
+        <li>
+          <a href="${pageContext.request.contextPath}/${name}">
+            Page: <c:out value="${name}" />
+          </a>
+        </li>
+      </c:forTokens>
     </ul>
   </div>
 </div>
